修改实体类，整理表与表的关系。

+++ src/main/java/com/zsk/dao/CategoryDao.java
@@ -3,6 +3,6 @@ package com.zsk.dao;
 import com.zsk.pojo.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryDao extends JpaRepository<Category,Long> {
+public interface CategoryDao extends JpaRepository<Category,Integer> {
 
 }
